Properly handle exit codes during set extraction
This makes sysupgrade fail early when unpacking a set fails, as should
have always been the case.

The fact that I had to do a one-off change like this makes me dubious
of the reliability of this script given that failures should always
abort.  I have not audited the whole script for more cases like this.
I see the script also doesn't carry "set -e", but IIRC that's
intentional because of the surprising behavior of this flag which
could cause even more obscure issues than this if used.

Fixes #5.